Move from override pkg to manifest pkg
